Replacing old state machine code
